requirements.txt
GitPython==2.1.3
giturlparse.py==0.0.5
requests==2.13.0
six==1.10.0
